继续美化代码：完善渲染系统和系统函数的变量名和函数名
- 美化03_rendering.c中的渲染着色器管理相关函数
- 添加渲染材质属性和着色器处理的语义化定义
- 完善渲染系统变量名的语义化替换
- 更新系统函数的宏定义和文档注释

+++ TaleWorlds.Native/src/01_initialization.c
@@ -28936,7 +28936,7 @@ void SystemFloatingPointProcessor(long long resourceManagerPointer,float floatVa
   if (isSystemConfigured) {
     scaleFactorX = (float)CalculateExponentialBase2(SystemNodeManagerPointer,contextParameter,AdditionalParameter,ConfigurationFlag,InvalidHandleValue);
     if (*(char *)(SystemResourceManager + 0x22d) == '\0') {
-      SystemMemoryPointer = *(long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8);
+      SystemMemoryPointer = *(uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8);
       if ((*(int *)(SystemMemoryPointer + 0x48) < SystemDataValuePrimary) &&
          (CheckSystemDataAvailability(&SystemDataValuePrimary), SystemDataValuePrimary == -1)) {
         SystemDataValueSecondary = ConfigurationDataPointer;
@@ -39605,7 +39605,7 @@ void ProcessSystemResourceSemaphore(long long SystemResourceManager,long long *C
   
   ConfigurationMask = (void*)0xfffffffffffffffe;
   (**(code **)(*(long long *)*ConfigurationDataPointer + 0x78))();
-  ThreadLocalStorageIndex = *(uint *)(*(long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8)
+  ThreadLocalStorageIndex = *(uint *)(*(uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8)
                    + 0xc);
   if (ThreadLocalStorageIndex == 0xffffffff) {
     ResourceOffset = GetResourceOffsetHandler(SystemResourceManager + 0x78);
@@ -39692,7 +39692,7 @@ void ProcessSystemResourceWithIntParameter(long long SystemResourceManager,void*
   long long ResourceOffset;
   long long BufferPointer;
   
-  ThreadLocalStorageIndex = *(uint *)(*(long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8)
+  ThreadLocalStorageIndex = *(uint *)(*(uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8)
                    + 0xc);
   if (ThreadLocalStorageIndex == 0xffffffff) {
     ResourceOffset = GetResourceOffsetHandler(SystemResourceManager + 0x78);
@@ -42383,7 +42383,7 @@ void InitializeSystemResource(long long* SystemResourceManager)
   char SystemLockAcquiredFlag;
   
   if ((char)SystemResourceManager[0xb] != '\0') {
-    PrimaryResourceHandle = (long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8);
+    PrimaryResourceHandle = (uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8);
     do {
       SystemOperationResult = (**(code **)(*SystemResourceManager + 0x20))(SystemResourceManager,1);
       if (validationStatusFlag == '\0') {
@@ -64481,7 +64481,7 @@ void* * GetSystemResourceConfigurationPointer(long long SystemResourceManager,vo
   if (*(long long *)(SystemResourceManager + 0x1b8) != 0) {
     return (void* *)(*(long long *)(SystemResourceManager + 0x1b8) + 0x10);
   }
-  if (*(int *)(*(long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8) +
+  if (*(int *)(*(uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8) +
               0x48) < SystemDataValue5) {
     CheckSystemDataAvailability(&SystemDataValue5);
     if (SystemDataValue5 == -1) {
@@ -67978,7 +67978,7 @@ SystemResourceFinalize:
   case 0xd:
     if (0 < SystemOperationResult) {
       SystemThreadHandleSecondary = 0x48;
-      SystemThreadHandle5 = *(long long *)((long long)ThreadLocalStorageBasePointer + (ulong long)__tls_index * 8);
+      SystemThreadHandle5 = *(uint64_t *)((uint64_t)ThreadLocalStorageBasePointer + (uint64_t)__tls_index * 8);
       MatrixElement28 = fRam0000000180d49154;
       do {
         SystemThreadHandleSecondary = *(long long *)(ConfigDataHandle + 0x68);
